feat(fe): 운영진용 모임 상세 페이지 구현 - 파일 분리
